Reworked some of the Hits pager logic and ESSearch behaviour to fix issue with hits results parsing.

+++ src/main/java/net/wedjaa/elasticparser/pager/ESAggregationPager.java
@@ -78,7 +78,7 @@ public class ESAggregationPager implements ESResultsPager {
 	}
 
 	@Override
-	public int current_hit_idx() {
+	public long current_hit_idx() {
 		return current_idx;
 	}
