+++ src/main/java/net/wedjaa/elasticparser/pager/ESEmptyPager.java
@@ -52,7 +52,7 @@ public class ESEmptyPager implements ESResultsPager {
 	}
 
 	@Override
-	public int current_hit_idx() {
+	public long current_hit_idx() {
 		return 0;
 	}
