+++ src/main/java/net/wedjaa/elasticparser/pager/ESFacetsPager.java
@@ -79,7 +79,7 @@ public class ESFacetsPager implements ESResultsPager {
 	}
 
 	@Override
-	public int current_hit_idx() {
+	public long current_hit_idx() {
 		return current_idx;
 	}
