+++ src/main/java/net/wedjaa/elasticparser/pager/ESResultsPager.java
@@ -31,7 +31,7 @@ public interface ESResultsPager {
 	public boolean done();
 	public boolean hit_available();
 	public void set_page_size(int page_size);
-	public int current_hit_idx();
+	public long current_hit_idx();
 	public int page_size();
 	public int next_page();
 	public String get_query();
